+++ src/App.jsx
@@ -1,6 +1,6 @@
 import { ControlContextProvider } from './context/ControlContext';
 import { ChatProvider } from './context/ChatContext';
-import { MessageProvider } from './context/MessageContext';
+
 import ControlPage from './pages/ControlPage/ControlPage';
 
 
@@ -8,9 +8,9 @@ const App = () => {
   return (
     <ControlContextProvider>
       <ChatProvider>
-        <MessageProvider>
+
           <ControlPage />
-        </MessageProvider>
+
       </ChatProvider>
     </ControlContextProvider>
   );
